Revert "GH-43902: [Java] Support for Long memory addresses (#43903)"
This reverts commit 6a4e804f238a6b8d007d5065a3f5b6d5c9a521e8.

+++ performance/src/main/java/org/apache/arrow/memory/AllocatorBenchmarks.java
@@ -57,9 +57,9 @@ public class AllocatorBenchmarks {
   @BenchmarkMode(Mode.AverageTime)
   @OutputTimeUnit(TimeUnit.MICROSECONDS)
   public void segmentRoundingPolicyBenchmark() {
-    final long bufferSize = 1024L;
+    final int bufferSize = 1024;
     final int numBuffers = 1024;
-    final long segmentSize = 1024L;
+    final int segmentSize = 1024;
 
     RoundingPolicy policy = new SegmentRoundingPolicy(segmentSize);
     try (RootAllocator allocator =
